feat(ssh): add SSH connection testing library and validation
Introduce a new `dodosh` library to handle SSH connections and authentication.
Use the library to validate SSH credentials before creating a server record.
This prevents creating servers with invalid SSH credentials and provides early feedback to users.

- Add `dodosh` crate with `test_connection` function
- Integrate SSH testing in `ServerService::create_remote_server`
- Update port fields from `i64` to `u16` for correctness
- Centralize workspace dependencies in root `Cargo.toml`

+++ backend/src/routes/create_server.rs
@@ -61,7 +61,7 @@ pub enum CreateServerRequest {
     Remote {
         name: String,
         hostname: String,
-        port: i64,
+        port: u16,
         auth: SshAuthRequest,
     },
 }
@@ -103,7 +103,7 @@ pub struct CreateServerResponse {
     #[serde(rename = "serverType")]
     pub server_type: types::ServerType,
     pub hostname: String,
-    pub port: Option<i64>,
+    pub port: Option<u16>,
 }
 
 #[utoipa::path(
